+++ scripts/create-changelog.js
@@ -1,0 +1,37 @@
+const changelog = require('../CHANGELOG.json');
+const fs = require('fs');
+
+if (changelog.versions && changelog.versions.length > 0) {
+  const markdown = [];
+
+  markdown.push(`# Releases`);
+  markdown.push(``);
+
+  // Loop over all the change log versions
+  for (const entry of changelog.versions) {
+    markdown.push(`## ${entry.version}`);
+    markdown.push(``);
+    // Check if entry contains change information
+    if (entry.changes) {
+      // Loop over all change types
+      for (const changeName in entry.changes) {
+        const typeChanges = entry.changes[changeName];
+        if (typeChanges.length > 0) {
+          let name = changeName === "new" ? "new control(s)" : changeName;
+          markdown.push(`**${name.charAt(0).toUpperCase() + name.slice(1)}**`);
+          markdown.push(``);
+
+          // Add each change text
+          for (const msg of typeChanges) {
+            markdown.push(`- ${msg}`);
+          }
+          markdown.push(``);
+        }
+      }
+    }
+  }
+
+  if (markdown.length > 2) {
+    fs.writeFileSync('CHANGELOG.md', markdown.join('\n'));
+  }
+}