+++ src/webparts/controlsTest/components/ControlsTest.tsx
@@ -226,7 +226,7 @@ export default class ControlsTest extends React.Component<IControlsTestProps, IC
                   // disabledTermIds={["943fd9f0-3d7c-415c-9192-93c0e54573fb", "0e415292-cce5-44ac-87c7-ef99dd1f01f4"]}
                   disabledTermIds={["943fd9f0-3d7c-415c-9192-93c0e54573fb", "73d18756-20af-41de-808c-2a1e21851e44", "0e415292-cce5-44ac-87c7-ef99dd1f01f4"]}
                   // disabledTermIds={["cd6f6d3c-672d-4244-9320-c1e64cc0626f", "0e415292-cce5-44ac-87c7-ef99dd1f01f4"]}
-                  disableChildrenOfDisabledParents={true}
+                  // disableChildrenOfDisabledParents={true}
                   panelTitle="Select Term"
                   label="Taxonomy Picker"
                   context={this.props.context}