+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ## 1.4.0
 
-**New Controls**
+**New control(s)**
 
 - `SecurityTrimmedControl` control got added [#74](https://github.com/SharePoint/sp-dev-fx-controls-react/issues/74)
 
@@ -23,7 +23,7 @@
 
 ## 1.3.0
 
-**New Controls**
+**Enhancements**
 
 - `TaxonomyPicker` control got added [#22](https://github.com/SharePoint/sp-dev-fx-controls-react/issues/22) [#63](https://github.com/SharePoint/sp-dev-fx-controls-react/issues/63) [#64](https://github.com/SharePoint/sp-dev-fx-controls-react/issues/64)
 - `ListPicker` control got added [#34](https://github.com/SharePoint/sp-dev-fx-controls-react/issues/34)
@@ -69,7 +69,7 @@
 
 ## 1.2.0
 
-**New controls**
+**New control(s)**
 
 - Field controls are added to the project
 - `IFrameDialog` was added to the project
@@ -80,31 +80,56 @@
 
 ## 1.1.3
 
-- `FileTypeIcon` icon fixed. This control should now also work in SPFx extensions.
+**Fixes**
+
+- `FileTypeIcon` icon fixed where it did not render an icon. This control should now works in SPFx extensions.
 
 ## 1.1.2
 
-- Fix for `WebPartTitle` control to inherit color
+**Enhancements**
+
 - Improved telemetry with some object checks
 
+**Fixes**
+
+- Fix for `WebPartTitle` control to inherit color
+
 ## 1.1.1
+
+**Enhancements**
 
 - Removed operation name from telemetry
 
 ## 1.1.0
 
+**Enhancements**
+
 - Telemetry added
 
 ## 1.0.0
-- **New control**: WebPartTitle control got added.
-- **Enhancement**: ListView control got extended with the ability to specify a set of preselected items.
+
+**New control(s)**
+
+- `WebPartTitle` control got added
+
+**Enhancements**
+
+- ListView control got extended with the ability to specify a set of preselected items.
 
 ## Beta 1.0.0-beta.8
-- **Bug fix**: bug fix for the `ListView` control when selection is used in combination with `setState`.
+
+**Fixes**
+
+- Fix for the `ListView` control when selection is used in combination with `setState`.
 
 ## Beta 1.0.0-beta.7
-**Added**
+
+**New control(s)**
+
 - Grouping functionality added to the `ListView` control
 
 ## Beta 1.0.0-beta.6
+
+**New control(s)**
+
 - Initial release